define some variable in middleware for evaluation (kafka bottleneck)

+++ database/migrations/2021_08_29_065152_create_process_logs_table.php
@@ -21,7 +21,7 @@ class CreateProcessLogsTable extends Migration
             $table->integer('task_count');
             $table->integer('result_count');
             $table->integer('success_percent');
-            $table->integer('avg_processing_duration');
+            $table->float('avg_processing_duration');
             $table->timestamps();
 
         });
